Move Role-Strategy security magic to a separate script
